% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{Bio_OnescRNAseq}
\alias{Bio_OnescRNAseq}
\title{process one scRNAseq dataset to generate biological metadata}
\usage{
Bio_OnescRNAseq(scdata, nHVGs = 1000, nPCs = 10, PCind = 1,
  organism = "mmusculus")
}
\arguments{
\item{scdata}{a SingleCellExperiment object;}

\item{nHVGs}{integer; the number of highly variable genes (default: 1000)}

\item{nPCs}{integer: the number of principal components (default: 10)}

\item{PCind}{integer; which principal component for exploring biological featues (default: 1; the first principal component will be used to find genes highly correlated with PCA 1); PCind should be less than nPC}

\item{organism}{a character string; the organism of single cell RNAseq datasets; if supported by WebGestaltR, functional enrichment analysis will be performed (defeault: mmusculus)}
}
\value{
a new SingleCellExperiment object by adding features into the the elementMetadata and metadata slots of the input object:
	                               
                                elementMetadata; hvg: a dataframe containing mean, variance and z-score for dispersion 
                                                 genevar_by_tounts:  variance explained by the log-transformed counts 
                                                 genevar_by_features: variance explained by the log-transformed features 
                                                 genevar_by_Mt: variance explained by the log-transformed mitochondrial counts
                                                 genevar_by_rRNA: variance explained by the log-transformed rRNA counts
                                metadata: A list containing other metadata, including
                                                 pc1genes: a dataframe containing the genes highly correlated with the 1st (default) or the PCind principal component 
                                                 pc1Pathway: a dataframe containing the pathways enriched in pc1 (default) or the PCind genes
                                                 hvgPathway: a dataframe containing the pathways enriched in top n (default:1000) highly variable genes
}
\description{
Generate metadata (HVGs, PC-related genes, variance explained by technical features) for one single cell RNAseq represented by gene-cell count table
}
\examples{
library(scRNABatchQC)
sce<-Tech_OnescRNAseq(inputfile="https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz")
sce<-Bio_OnescRNAseq(sce)
head(sce@elementMetadata)
sce@metadata$hvgPathway
}
