% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{Combine_scRNAseq}
\alias{Combine_scRNAseq}
\title{Combine scRNAseq datasets to one dataset and run the analysis on the combined dataset}
\usage{
Combine_scRNAseq(sces, nHVGs = 1000, nPCs = 10, logFC = 1, FDR = 0.01,
  sampleRatio = 1, organism = "mmusculus")
}
\arguments{
\item{sces}{a list of SingleCellExperiment objects; each object corresponds to one single cell RNAseq dataset}

\item{nHVGs}{integer; the number of highly variable genes (default:1000)}

\item{nPCs}{integer: the number of principal components (default:10)}

\item{logFC}{float; log fold change cutoff to select differentially expressed genes (default: 1)}

\item{FDR}{float; FDR cutoff to select differentially expressed genes (default: 0.01)}

\item{sampleRatio}{float; the ratio of cells sampled from each dataset to examine the expression similarity(default: 1)}

\item{organism;}{a character giving the organism of single cell RNAseq datasets (default: hsapiens)}
}
\value{
a SingleCellExperiment object with several slots:
	                               assays; ShallowSimpleListAssays object containing one sparse matrix:  logcounts (log-transformed normalized counts)
                                elementMetadata; A DataFrame containining metadata for each gene, including
                                                 hvg: a dataframe containing mean, variance and z-score for dispersion 
                                colData; A DataFrame containing metadata for each cell, including
                                                 condition: names where the cell come from
                                metadata: A list containing other metadata, including
                                                  reducedDims: a list containing two types of reduced dimensions: PCA and tSNE
	                                                 diffFC: differential expressed genes and pathways from pairwise comparison between any two datasets
                                                  other metadata including nHVGs,nPC, logFC, FDR, sampleRatio
}
\description{
Combine and compare multiple single cell RNAseq datasets; 1) combine into one dataset; 2)find highly variable genes, reduce dimensions using PCA and tSNE for this combined dataset; 3) pairwise comparsion between any two single cell RNAseq datasets to find the top differentially expressed genes
}
\examples{
library(scRNABatchQC)
sces<-process_scRNAseq(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
scmerge <- Combine_scRNAseq(sces)
logcounts(scmerge)[1:5,1:5]
head(scmerge@elementMetadata$hvg)
summary(scmerge@metadata$reducedDims$PCA)
visualize PCA results
plot(scmerge@metadata$reducedDims$PCA$x[,1:2],pch=16,col=as.factor(scmerge@colData$condition),xlab="PCA1",ylab="PCA2")
visualize tSNE results
plot(scmerge@metadata$reducedDims$tSNE,pch=16,col=as.factor(scmerge@colData$condition),xlab="tSNE1",ylab="tSNE2")
scmerge@metadata$diffFC
}
