% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{generateReport}
\alias{generateReport}
\title{generate a QC report in a html file}
\usage{
generateReport(sces, scesMerge, outputFile = "report.html")
}
\arguments{
\item{sces}{a list of SingleCellExperiment Objects; each object containing quality meta data for one single cell RNAseq dataset}

\item{scesMerge}{a SingleCellExperiment object generated by combining and comparing multiple single cell RNAseq datasets}

\item{outputFile}{the name of the output file (default: report.html)}
}
\value{
a list of SingleCellExperiment
}
\description{
Generate a QC report by comparing technical and biological features across multiple scRNA-seq datasets
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
scesMerge <- Combine_scRNAseq(sces)
generateReport(sces,scesMerge)
}
