% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotAlltSNE}
\alias{plotAlltSNE}
\title{plot tSNE for the combined dataset}
\usage{
plotAlltSNE(scesMerge, scolors = NULL, pointSize = 0.8)
}
\arguments{
\item{scesMerge}{SingleCellExperiment object; this object contains the combined data and reduced dimensions using PCA and tSNE}

\item{scolors}{a vector of color}

\item{pointSize}{integer; size of point}
}
\description{
tSNE plot for the combined datasets
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                   "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
scesMerge<-Combine_scRNAseq(sces)
plotAlltSNE(scesMerge)
}
