% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotAveCountVSdetectRate}
\alias{plotAveCountVSdetectRate}
\title{plotAveCountVSdetectRate}
\usage{
plotAveCountVSdetectRate(sces, scolors = 1:length(sces), lineSize = 1)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset}

\item{scolors}{a vector of integer; the color of each dataset (default: 1:length(sces))}

\item{lineSize}{integer; the line size of the plot  (default: 1)}
}
\description{
plot the gene detection rates vs. the average gene expression level in the datasets
}
\examples{
library(scRNABatchQC)
sces<-prepareSCRNADataSet(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                      "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"),
                          organism="mmusculus")
#plot the gene detection rates for all datasets
plotAveCountVSdetectRate(sces)
#plot the gene detection rates for the first dataset
plotAveCountVSdetectRate(sces[1])
}
