% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDiffgenes}
\alias{plotDiffgenes}
\title{plotDiffgenes}
\usage{
plotDiffgenes(scesmerge, margins = c(5, 5), keysize = 1,
  col = bluered(75), ...)
}
\arguments{
\item{scesmerge}{a SingleCellExperiment object; this object contains the combined datasets, pairwise comparison results and reduced dimensions using PCA and tSNE}

\item{margins}{margins for heatmap.2}

\item{keysize}{integer for heatmap.2}

\item{col}{color for heatmap.2}

\item{...}{parameters passing to heatmap.2}
}
\description{
plot the differentially expressed genes in any pairwise comparison
}
\examples{
library(scRNABatchQC)
sces<-prepareSCRNADataSet(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                      "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"),
                          organism="mmusculus")
scesMerge<-preparePCATSNEData(sces,organism="mmusculus")
#plot the average count distribution for all datasets
plotDiffgenes(scesMerge)
}
