% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotGeneCountDistribution}
\alias{plotGeneCountDistribution}
\title{plot the gene count distribution}
\usage{
plotGeneCountDistribution(
  sces,
  scolors = 1:length(sces),
  ngenes = 500,
  lineSize = 1
)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset; (results from \code{\link{Process_scRNAseq}} )}

\item{scolors}{a vector of integer; the color of each dataset (default: 1:length(sces))}

\item{ngenes}{integer; the number of highly expressed genes to plot the distribution (default:500)}

\item{lineSize}{integer; the line size of the plot  (default: 1)}
}
\description{
plot the gene count distribution of top n (default:500) highly expressed genes
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputfiles=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                    "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
  
#plot the average count distribution for all datasets
plotGeneCountDistribution(sces)
#plot the average count distribution for the first dataset
plotGeneCountDistribution(sces[1])
}
\seealso{
\code{\link{Process_scRNAseq}}
}
