% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotHVGsPathways}
\alias{plotHVGsPathways}
\title{plot pathways enriched in HVGs}
\usage{
plotHVGsPathways(sces, margins = c(5, 10), keysize = 1,
  col = colorpanel(75, low = "white", high = "red"), ...)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset}

\item{margins}{margins for heatmap.2 (default: c(5,10))}

\item{keysize}{integer for heatmap.2  (default: 1)}

\item{col}{color for heatmap.2  (default: colorpanel(75,low="white",high="red"))}

\item{...}{parameters passing to heatmap.2}
}
\value{
a matrix containing the -log10 pvalue of enriched pathways in each dataset
}
\description{
plot the pathways enriched in the highly variable genes across single cell RNAseq datasets
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputfiles=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                    "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
plotHVGsPathways(sces)
}
