% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotPCgenes}
\alias{plotPCgenes}
\title{plot PC-related genes}
\usage{
plotPCgenes(sces, margins = c(5, 5), keysize = 1, col = bluered(75), ...)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset}
}
\value{
a matrix containing the log fold change of genes highly associated with a certain principle component
}
\description{
plot the genes highly associated with a certain principle component
}
\examples{
library(scRNABatchQC)
sces<-prepareSCRNADataSet(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"),organism="mmusculus")
#plot the average count distribution for all datasets
plotPCgenes(sces)
}
