% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotPCPathways}
\alias{plotPCPathways}
\title{plot PC-related pathways}
\usage{
plotPCPathways(sces, margins = c(5, 10), keysize = 1, col = colorpanel(75,
  low = "white", high = "red"), ...)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset}
}
\value{
a matrix containing the -log10 pvalue of enriched pathways in each dataset
}
\description{
plot the pathways enriched in the PC genes across single cell RNAseq datasets
}
\examples{
library(scRNABatchQC)
sces<-prepareSCRNADataSet(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"),organism="mmusculus")
#plot the average count distribution for all datasets
plotPCPathways(sces)
}
