% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotVarianceExplained}
\alias{plotVarianceExplained}
\title{plot variance explained by one specific feature}
\usage{
plotVarianceExplained(sces, feature = c("genevar_by_counts",
  "genevar_by_features", "genevar_by_Mt", "genevar_by_rRNA"),
  scolors = 1:length(sces), lineSize = 1)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset}

\item{feature}{a character; the explanatory variable to plot; feature should be genevar_by_counts, genevar_by_features, genevar_by_Mt, or genevar_by_rRNA; (default: genevar_by_counts)}

\item{scolors}{a vector of integer; the color of each dataset (default: 1:length(sces))}

\item{lineSize}{integer; the line size of the fitted mean-variance trend curve  (default: 1)}
}
\description{
plot the variances explained by the explanatory variable
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputfiles=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                    "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
  
#plot the variance explained by the total_counts 
plotVarianceExplained(sces)
#plot the variance explained by the total number of genes
plotVarianceExplained(sces,feature="genevar_by_features")
}
