% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotVarianceTrend}
\alias{plotVarianceTrend}
\title{plot Mean~variance trend}
\usage{
plotVarianceTrend(sces, scolors = 1:length(sces), pointSize = 0.8,
  lineSize = 1)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset; (results from \code{\link{Process_scRNAseq}} )}

\item{scolors}{a vector of integer; the color of each dataset (default: 1:length(sces))}

\item{lineSize}{integer; the line size of the fitted mean-variance trend curve  (default: 1)}

\item{pointSize;}{the point size of each dot representing each gene}
}
\description{
plot the gene mean vs. variance trend
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputfiles=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                    "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
  
#plot the mean~variance trend for all datasets
plotVarianceTrend(sces)
#plot the mean~variance trend for the first dataset
plotVarianceTrend(sces[1])
}
\seealso{
\code{\link{Process_scRNAseq}}
}
