% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSCRNADataSet.R
\name{preparePCATSNEData}
\alias{preparePCATSNEData}
\title{preparePCATSNEData}
\usage{
preparePCATSNEData(sces, nHVGs = 1000, nPC = 10)
}
\arguments{
\item{sces}{a named list of scRNA data}

\item{nHVGs}{integer; the number of highly variable genes (default:1000)}

\item{nPC}{integer: the number of principal components (default:10)}
}
\value{
a list with PCA and TSNE data
}
\description{
generate the PCA and tSNE data for the combined samples.
}
\examples{
library(scRNABatchQC)
sces <- prepareSCRNADataSet(inputfiles=c("data1.csv","data2.csv"))
sceall <- preparePCATSNEData(sces)
}
