% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSCRNADataSet.R
\name{prepareSCRNADataSet}
\alias{prepareSCRNADataSet}
\title{prepareSCRNADataSet}
\usage{
prepareSCRNADataSet(inputfiles, samplenames = NULL,
  organism = c("hsapiens", "mmusculus"), sampleRatio = 1,
  nHVGs = 1000, sf = 10000)
}
\arguments{
\item{sampleRatio}{float; the ratio of cells sampled from each experiment to examine the expression similarity(default: 1, use the full dataset without sampling)}

\item{nHVGs}{integer; the number of highly variable genes (default: 1000)}

\item{sf}{integer; Scale factor  (default: 10000)}

\item{inputfiles;}{a character vector giving the input file name (or a URL starting http://, file://, etc.) of gene-by-cell count matrices, the rowname should be gene symbol; each file should be regular delimited file;  Compressed files ending .gz and .bz2 are supported}
}
\value{
a named list containing each scRNA data sets,
}
\description{
Generate statistics information for multiple single-cell RNA-seq datasets represented by gene-count matrices
}
\examples{
library(scRNABatchQC)
sces<-prepareSCRNADataSet(inputfiles=c("data1.csv","data2.csv"))
}
