% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{Process_OnescRNAseq}
\alias{Process_OnescRNAseq}
\title{process one scRNAseq dataset to generate QC metadata}
\usage{
Process_OnescRNAseq(
  input,
  sf = 10000,
  mincounts = 500,
  mingenes = 200,
  maxmito = 0.2,
  mtRNA = "^mt-|^MT-",
  rRNA = "^Rp[sl][[:digit:]]|^RP[SL][[:digit:]]",
  nHVGs = 1000,
  nPCs = 10,
  PCind = 1,
  organism = "mmusculus",
  chunk.size = NULL
)
}
\arguments{
\item{input}{string of file or path name, a SingleCellExperiment or Seurat v3 object; \cr
input can be the file name (or a URL starting http://, file://, etc.) of gene-by-cell count matrix, the rowname should be gene symbol; the file should be regular delimited file;  Compressed files ending .gz and .bz2 are supported. \cr
input can be the path name, which contains the barcodes.tsv.gz, features.tsv.gz, and matrix.mtx.gz provided by 10X from CellRanger >=3.0 \cr
input can also be a SingleCellExperiment or Seurat v3 object}

\item{sf}{integer; Scale factor to normalize the single cell RNA-seq data (default: 10000)}

\item{mincounts}{integer; the cutoff of filtering the cell if the total number of counts in the cell less than the mincounts (default:500)}

\item{mingenes}{integer; the cutoff of filtering the cell if the total number of genes detected in the cell less than the mingenes (default: 200)}

\item{maxmito}{float; the cutoff of filtering the cell if the percentage of mtRNA reads in the cell larger than the minmito; (default: 0.2);}

\item{mtRNA}{string; the pattern of genenames for mitochondrial encoded RNAs ; (default: "^mt-|^MT-", the default is mtRNA genenames in human or mouse); If not human or mouse, input the gene name pattern of mtRNA}

\item{rRNA}{string; the pattern of genenames for ribosomal proteins; (default: "^Rp[sl][[:digit:]]|^RP[SL][[:digit:]]", the default is ribosomal protein genenames in human or mouse); If not human or mouse, input the gene name pattern of ribosomal proteins}

\item{nHVGs}{integer; the number of highly variable genes (default: 1000)}

\item{nPCs}{integer: the number of principal components (default: 10)}

\item{PCind}{integer; which principal component for exploring biological featues (default: 1; the first principal component will be used to find genes highly correlated with PCA 1); PCind should be less than nPC}

\item{organism}{string; the organism of single cell RNAseq datasets; if supported by WebGestaltR, functional enrichment analysis will be performed (defeault: mmusculus)}

\item{chunk.size}{NULL or integer; default is NULL, suggesting data will be loaded into memory at one time, otherwise, the data will be loaded into memory by chunks with chunk.size}
}
\value{
a SingleCellExperiment object with several slots:
\itemize{
\item  {                            assays; ShallowSimpleListAssays object containing two sparse matrix: counts and logcounts }
\item  {                            rowRanges@elementMetadata; A DataFrame containining metadata for each gene, including 
           \itemize{  
                        \item {               ave.counts: the average counts  }
                        \item {               num.cells: the number of cells with the gene detected }
                        \item {               hvg: a dataframe containing mean, variance and z-score for dispersion }
                        \item {               genevar_by_tounts:  variance explained by the log-transformed counts }
                        \item {               genevar_by_features: variance explained by the log-transformed features }
                        \item {               genevar_by_Mt: variance explained by the log-transformed mitochondrial counts }
                        \item {               genevar_by_rRNA: variance explained by the log-transformed rRNA counts }
             }
        }
 \item {                             colData; A DataFrame containing metadata for each cell, including
              \itemize{
                          \item{                log10_total_counts: the number of counts in each cell in log10 transformed }
                          \item {               log10_total_features: the number of genes detected in each cell in log10 transformed }
                          \item {               log10_total_counts_Mt: the number of mitochondrial counts in each cell in log10 transformed }
                          \item {               log10_total_counts_rRNA: the number of rRNA counts in each cell in log10 transformed  }
               }
}
 \item {                               metadata: A list containing other metadata, including
               \itemize{
                             \item {         rawmeta: a list of metadata for genes and cells of raw data before filtering, including
                                   \itemize{ 
                                         \item {        sf: normalization factor }
                                         \item {        ngenes: the number of genes }
                                         \item {        ncells: the number of cells }
                                         \item {        CellData: a dataframe containing metadata for each cell before filtering, includingtotal_counts,total_features,total_counts_Mt,total_counts_rRNA, pct_counts_rRNA (perentage of rRNA counts), pct_counts_Mt (percentage of mtRNA counts), libsize.drop(cell is filtered by library size), feature.drop (cell is filtered by the number of detected genes), mito.drop (cell is filtered by the mtRNA counts), is.drop (cell is filtered by either of library size, the number of genes or the mtRNA reads) }
                                         \item {        GeneData: a list containing the gene filter information (gene.keep, gene is filtered since none of the cells detect theg gene)    }
                                        \item {        Cutoff: a list containing the cutoff values (count, gene, mito) for filtering cells                                                }   
                                        }
                                 }
                        \item {           pc1genes: a dataframe containing the genes highly correlated with the 1st (default) or the PCind principal component }
                        \item {           pc1Pathway: a dataframe containing the pathways enriched in pc1 (default) or the PCind genes  }
                        \item {           hvgPathway: a dataframe containing the pathways enriched in top n (default:1000) highly variable genes }
                       }                                            
    }
 }
}
\description{
Generate technical and biological metadata for one single cell RNAseq dataset
}
\examples{
library(scRNABatchQC)
sce<-Process_OnescRNAseq(input="https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz")
head(sce@rowRanges@elementMetadata)
head(colData(sce))
counts(sce)[1:5,1:5]
logcounts(sce)[1:5,1:5]
sce@metadata$rawmeta$ngenes
head(sce@metadata$rawmeta$CellData)
head(sce@metadata$pc1Pathway)
sces=list(sce=sce)
plotDensity(sces)
}
\seealso{
\code{\link{Process_scRNAseq}}, \code{\link{Combine_scRNAseq}}
}
