% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{scRNABatchQC}
\alias{scRNABatchQC}
\title{scRNABatchQC}
\usage{
scRNABatchQC(inputfiles, samplenames = NULL, organism = c("hsapiens",
  "mmusculus"), outputFile = "report.html", sampleRatio = 1,
  cache = FALSE, nHVGs = 1000, nPC = 10, sf = 10000, logFC = 1,
  FDR = 0.01)
}
\arguments{
\item{sampleRatio}{float; the ratio of cells sampled from each experiment to examine the expression similarity(default: 1)}

\item{nHVGs}{integer; the number of highly variable genes (default: 1000)}

\item{nPC}{integer; the number of principle components (default: 10)}

\item{sf}{integer; Scale factor  (default: 10000)}

\item{logFC}{float; log2 fold change cutoff to select differentially expressed genes  (default: 1)}

\item{FDR}{float; FDR cutoff to select differentially expressed genes (default:0.01)}

\item{inputfiles;}{a character vector giving the file names containing gene-cell count matrices; each file should be delimited, either in tsv or csv format}

\item{samplenames;}{the names of experiments, if it is NULL(default), the names are set to S1, S2....}

\item{organism;}{a character string giving the organism, either hsapiens or mmusculus; (default: "hsapiens")}

\item{outputFile;}{a character string giving the output File name (default: report.html)}
}
\description{
Compare multiple scRNA-seq datasets simultaneously on numerous technical and biological features
}
\examples{
library(scRNABatchQC)  
#Check the quality of two single cell RNA-seq datasets from human
scRNABatchQC(inputfiles=c("data1.csv","data2.csv"))  

}
