% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{generateReport}
\alias{generateReport}
\title{generate a QC report in a html file}
\usage{
generateReport(sces, scesMerge, outputFile = "report.html",
  lineSize = 1, pointSize = 0.8)
}
\arguments{
\item{sces}{a list of SingleCellExperiment Objects; each object containing quality meta data for one single cell RNAseq dataset (results from \code{\link{Process_scRNAseq}})}

\item{scesMerge}{a SingleCellExperiment object generated by combining and comparing multiple single cell RNAseq datasets (results from \code{\link{Combine_scRNAseq}})}

\item{outputFile}{the name of the output file (default: report.html)}

\item{lineSize}{float; the line size of figures in the generated report (default: 1)}

\item{pointSize}{float; the point size of figures in the generated report (default: 0.8)}
}
\description{
Generate a QC report by comparing technical and biological features across multiple scRNA-seq datasets
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputs=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
scesMerge <- Combine_scRNAseq(sces)
generateReport(sces,scesMerge)
}
\seealso{
\code{\link{Process_scRNAseq}} \code{\link{Combine_scRNAseq}}
}
