% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotAllPCA}
\alias{plotAllPCA}
\title{plot PCA for the combined dataset}
\usage{
plotAllPCA(scesMerge, scolors = NULL, pointSize = 0.8)
}
\arguments{
\item{scesMerge}{SingleCellExperiment object; this object contains the combined data and reduced dimensions using PCA and tSNE; (results from \code{\link{Combine_scRNAseq}})}

\item{scolors}{a vector of color}

\item{pointSize}{integer; size of point}
}
\description{
PCA plot for the combined dataset
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                   "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
scesMerge <- Combine_scRNAseq(sces)
plotAllPCA(scesMerge)
}
