% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotHVGs}
\alias{plotHVGs}
\title{plotHVGs}
\usage{
plotHVGs(sces, margins = c(5, 5), keysize = 0.6, col = bluered(75),
  ...)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset}

\item{margins}{margins for heatmap.2}

\item{keysize}{integer for heatmap.2}

\item{col}{color for heatmap.2}

\item{...}{parameters passing to heatmap.2}
}
\value{
a matrix containing the z score of dispersion of highly variable genes in each dataset
}
\description{
plot the highly variable genes across single cell RNAseq datasets
}
\examples{
library(scRNABatchQC)
sces<-prepareSCRNADataSet(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz",
                                      "https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"),
                          organism="mmusculus")
#plot the average count distribution for all datasets
plotHVGs(sces)
}
