% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotSampleSimilarity}
\alias{plotSampleSimilarity}
\title{plot the Sample Similarity}
\usage{
plotSampleSimilarity(sces, ...)
}
\arguments{
\item{sces}{list; a list of SingleCellExperiment objects; each object containing QC metadata for each dataset}
}
\description{
plot the average expression similarity across single cell RNAseq datasets
}
\examples{
library(scRNABatchQC)
sces<-prepareSCRNADataSet(inputfile=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"),organism="mmusculus")
#plot the average count distribution for all datasets
plotSampleSimilarity(sces)
}
