% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSCRNAData.R
\name{prepareSCRNAData}
\alias{prepareSCRNAData}
\title{prepareSCRNAData}
\usage{
prepareSCRNAData(inputfile, organism = c("hsapiens", "mmusculus"),
  sampleRatio = 1, nHVGs = 1000, sf = 10000)
}
\arguments{
\item{sampleRatio}{float; the ratio of cells sampled from each experiment to examine the expression similarity(default: 1)}

\item{nHVGs}{integer; the number of highly variable genes (default: 1000)}

\item{sf}{integer; Scale factor  (default: 10000)}

\item{inputfile;}{the input file name (or a URL starting http://, file://, etc.) of gene-by-cell count matrix, the rowname should be gene symbol; the file should be regular delimited file;  Compressed files ending .gz and .bz2 are supported}
}
\value{
a named list containing scRNA data, 
                                hvggenes:high variance genes, 
                                pc1genes: genes in first principle component,
}
\description{
Generate statistics information from one single cell RNAseq represented by gene-cell count table
}
\examples{
library(scRNABatchQC)
sce<-prepareSCRNAData(inputfile="data1.csv")
}
