% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{Process_scRNAseq}
\alias{Process_scRNAseq}
\title{process scRNAseq datasets one by one to generate QC metadata}
\usage{
Process_scRNAseq(
  inputs,
  names = NULL,
  nHVGs = 1000,
  nPCs = 10,
  sf = 10000,
  mincounts = 500,
  mingenes = 200,
  maxmito = 0.2,
  PCind = 1,
  mtRNA = "^mt-|^MT-",
  rRNA = "^Rp[sl][[:digit:]]|^RP[SL][[:digit:]]",
  organism = "mmusculus",
  chunk.size = NULL
)
}
\arguments{
\item{inputs}{a string vector of file or path names, or a list of SingleCellExperiment or Seurat v3 objects; \cr
inputs can be a string vector of file names (or a URL starting http://, file://, etc.) of gene-by-cell count matrices, the rowname should be gene symbol; each file should be regular delimited file;  Compressed files ending .gz and .bz2 are supported. \cr
inputs can be a string vector of path names, each of which contains the barcodes.tsv.gz, features.tsv.gz, and matrix.mtx.gz provided by 10X from CellRanger >=3.0 \cr
inputs can also be a list of SingleCellExperiment or Seurat v3 objects}

\item{names}{string vector; giving the names of single-cell RNAseq datasets (default: NULL); names should have the same length of inputs; if NULL, the names are S1, S2...}

\item{nHVGs}{integer; the number of highly variable genes (default: 1000)}

\item{nPCs}{integer; the number of principal components (default: 10)}

\item{sf}{integer; Scale factor to normalize the data  (default: 10000)}

\item{mincounts}{integer; the cutoff of filtering the cell if the total number of counts in the cell less than the mincounts (default:500)}

\item{mingenes}{integer; the cutoff of filtering the cell if the total number of genes detected in the cell less than the mingenes (default: 200)}

\item{maxmito}{float; the cutoff of filtering the cell if the percentage of mtRNA reads in the cell larger than the minmito (default: 0.2)}

\item{PCind}{integer; which principal component for exploring biological featues (default: 1; the first principal component will be used to find genes highly correlated with PCA 1); PCind should be less than nPC}

\item{mtRNA}{string; the pattern of gene names for mitochondrial encoded RNAs ; (default: "^mt-|^MT-", the default is mtRNA gene names in human or mouse); If not human or mouse, give the gene name pattern of mtRNA}

\item{rRNA}{string; the pattern of gene names for ribosomal proteins; (default: "^Rp[sl][[:digit:]]|^RP[SL][[:digit:]]", the default is ribosomal protein gene names in human or mouse); If not human or mouse, give the gene name pattern of ribosomal proteins}

\item{organism}{string; the organism of single cell RNAseq datasets; if supported by WebGestaltR, functional enrichment analysis will be performed (defeault: mmusculus)}

\item{chunk.size}{NULL or integer; default is NULL, suggesting data will be loaded into memory at one time, otherwise, the data will be loaded into memory by chunks with chunk.size}
}
\value{
a list of SingleCellExperiment objects ;  \cr
 each SingleCellExperiment object containing metadata for one single cell RNAseq dataset;  \cr
        each SingleCellExperiment object containing several slots:
\itemize{
\item  {                        assays; ShallowSimpleListAssays object containing two sparse matrix: counts and logcounts }
\item  {                        rowRanges@elementMetadata; A DataFrame containining metadata for each gene, including 
           \itemize{  
                        \item {              ave.counts: the average counts  }
                        \item {              num.cells: the number of cells with the gene detected }
                        \item {              hvg: a dataframe containing mean, variance and z-score for dispersion }
                        \item {              genevar_by_tounts:  variance explained by the log-transformed counts }
                        \item {              genevar_by_features: variance explained by the log-transformed features }
                        \item {              genevar_by_Mt: variance explained by the log-transformed mitochondrial counts }
                        \item {              genevar_by_rRNA: variance explained by the log-transformed rRNA counts }
             }
        }
 \item {                        colData; A DataFrame containing metadata for each cell, including
              \itemize{
                          \item {                log10_total_counts: the number of counts in each cell in log10 transformed }
                          \item {               log10_total_features: the number of genes detected in each cell in log10 transformed }
                          \item {               log10_total_counts_Mt: the number of mitochondrial counts in each cell in log10 transformed }
                          \item {               log10_total_counts_rRNA: the number of rRNA counts in each cell in log10 transformed  }
               }
       }
 \item {                               metadata: A list containing other metadata, including
            \itemize{
                         \item {         rawmeta: a list of metadata for genes and cells of raw data before filtering, including
                                  \itemize{ 
                                         \item {        sf: normalization factor }
                                         \item {        ngenes: the number of genes }
                                         \item {        ncells: the number of cells }
                                         \item {        CellData: a dataframe containing metadata for each cell before filtering, includingtotal_counts,total_features,total_counts_Mt,total_counts_rRNA, pct_counts_rRNA (perentage of rRNA counts), pct_counts_Mt (percentage of mtRNA counts), libsize.drop(cell is filtered by library size), feature.drop (cell is filtered by the number of detected genes), mito.drop (cell is filtered by the mtRNA counts), is.drop (cell is filtered by either of library size, the number of genes or the mtRNA reads) }
                                         \item {        GeneData: a list containing the gene filter information (gene.keep, gene is filtered since none of the cells detect theg gene)    }
                                        \item {        Cutoff: a list containing the cutoff values (count, gene, mito) for filtering cells                                                }   
                                        }
                                 }
                        \item {           pc1genes: a dataframe containing the genes highly correlated with the 1st (default) or the PCind principal component }
                        \item {           pc1Pathway: a dataframe containing the pathways enriched in pc1 (default) or the PCind genes  }
                        \item {           hvgPathway: a dataframe containing the pathways enriched in top n (default:1000) highly variable genes }
                       }                                            
    }
 }
}
\description{
Generate technical and biological metadata for one or multiple single-cell RNAseq datasets;each dataset is processed one by one
}
\examples{
library(scRNABatchQC)
sces<-Process_scRNAseq(inputs=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
names(sces)
class(sces[[1]])
head(sces[[1]]@rowRanges@elementMetadata)
head(sces[[2]]@rowRanges@elementMetadata)
head(colData(sces[[1]]))
sces[[1]]@metadata$rawmeta$ngenes
head(sces[[1]]@metadata$rawmeta$CellData)
head(sces[[1]]@metadata$pc1Pathway)
plotDensity(sces,"total_counts")
plotVarianceTrend(sces)
plotPCPathways(sces)
}
\seealso{
\code{\link{Combine_scRNAseq}} , \code{\link{generateReport}}
}
