% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNABatchQC.R
\name{scRNABatchQC}
\alias{scRNABatchQC}
\title{scRNABatchQC}
\usage{
scRNABatchQC(inputfiles, names = NULL, nHVGs = 1000, nPCs = 10,
  sf = 10000, mincounts = 500, mingenes = 200, maxmito = 0.2,
  PCind = 1, mtRNA = "^mt-|^MT-",
  rRNA = "^Rp[sl][[:digit:]]|^RP[SL][[:digit:]]", sampleRatio = 1,
  logFC = 1, FDR = 0.01, organism = "mmusculus",
  outputFile = "report.html", lineSize = 1, pointSize = 0.8,
  chunk.size = NULL)
}
\arguments{
\item{inputfiles}{string vector; giving the input file names (or a URL starting http://, file://, etc.) of gene-by-cell count matrices, the rowname should be gene symbol; each file should be regular delimited file;  Compressed files ending .gz and .bz2 are supported}

\item{names}{string vector; giving names of each sample  (default: NULL); names should have the same length of inputfiles; if NULL, the names are S1, S2...}

\item{nHVGs}{integer; the number of highly variable genes (default: 1000)}

\item{nPCs}{integer; the number of principal components (default: 10)}

\item{sf}{integer; Scale factor to normalize the single cell RNA-seq data (default: 10000)}

\item{mincounts}{integer; the cutoff of filtering the cell if the total number of counts in the cell less than the mincounts (default:500)}

\item{mingenes}{integer; the cutoff of filtering the cell if the total number of genes detected in the cell less than the mingenes (default: 200)}

\item{maxmito}{float; the cutoff of filtering the cell if the percentage of mtRNA reads in the cell larger than the minmito (default: 0.2)}

\item{PCind}{integer; which principal component for exploring biological featues (default: 1; the first principal component will be used to find genes highly correlated with PCA 1); PCind should be less than nPC}

\item{mtRNA}{string; the pattern of genenames for mitochondrial encoded RNAs ; (default: "^mt-|^MT-", the default is mtRNA genenames in human or mouse); If not human or mouse, input the gene name pattern of mtRNA}

\item{rRNA}{string; the pattern of genenames for ribosomal proteins; (default: "^Rp[sl][[:digit:]]|^RP[SL][[:digit:]]", the default is ribosomal protein genenames in human or mouse); If not human or mouse, input the gene name pattern of ribosomal proteins}

\item{sampleRatio}{float; the ratio of cells sampled from each dataset to examine the expression similarity (default: 1)}

\item{logFC}{float; log fold change cutoff to select differentially expressed genes (default: 1)}

\item{FDR}{float; FDR cutoff to select differentially expressed genes (default: 0.01)}

\item{organism}{string; the organism of single cell RNAseq datasets; if supported by WebGestaltR, functional enrichment analysis will be performed (default: mmusculus);WebGestaltR supports 12 organisms, including athaliana, btaurus,celegans, cfamiliaris, drerio, sscrofa, dmelanogaster, ggallus, hsapiens, mmusculus, rnorvegicus, and scerevisiae.}

\item{outputFile}{string; the name of the output file (default: report.html)}

\item{lineSize}{float; the line size of figures in the report (default: 1)}

\item{pointSize}{float; the point size of figures in the report  (default: 0.8)}

\item{chunk.size}{NULL or integer; default is NULL, suggesting data will be loaded into memory at one time, otherwise, the data will be loaded into memory by chunks with chunk.size}
}
\value{
a list of SingleCellExperiment objects;
 \itemize{
 \item  {       sces: a list of SingleCellExperiment objects; each object contains technical and biological metadata for one scRNAseq dataset; see the output of  \code{\link{Process_scRNAseq} }}
 \item {        scesMerge: a SingleCellExperiment object containing bilogical metadata for the combined dataset and the pairwise difference across datasets; see the output of \code{\link{Combine_scRNAseq} }}
}
}
\description{
Compare multiple scRNA-seq datasets simultaneously on numerous technical and biological features
}
\examples{
library(scRNABatchQC)  
output<-scRNABatchQC(inputfiles=c("https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar1.csv.gz","https://github.com/liuqivandy/scRNABatchQC/raw/master/bioplar5.csv.gz"))
output$sces
output$scesMerge
plotDensity(output$sces, "total_counts")
output$sces[[1]]@metadata$hvgPathway
plotHVGs(output$sces)
output$scesMerge@metadata$diffFC$genes

}
\seealso{
\code{\link{Process_scRNAseq}}, \code{\link{Combine_scRNAseq}} , \code{\link{generateReport}}
}
